backend/functions/orchestrator/requirements.txt
functions-framework==3.5.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.20.2
flask==3.0.3
pandas==2.2.2
numpy==2.0.2
pyarrow==16.1.0
google-generativeai==0.7.2
firebase-admin==6.5.0
matplotlib==3.8.4
seaborn==0.13.2